Add PVC tests and move XMX8 tests to a new folder (#1347)
