+++ SYCL/Matrix/XMX8/element_wise_irreg_sum_rows.cpp
@@ -1,17 +1,19 @@
-//==-------- joint_matrix_bf16_vnni.cpp  - DPC++ joint_matrix---------------==//
+//==-------- element_wise_irreg_sum_rows.cpp  - DPC++ joint_matrix----- ----==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
-// REQUIRES: matrix
+// REQUIRES: matrix-xmx8
 
 // RUN: %clangxx -fsycl %s -o %t.out
 // RUN: %CPU_RUN_PLACEHOLDER %t.out
 // RUN: %GPU_RUN_PLACEHOLDER %t.out
 
-// XFAIL: gpu
+// this code calculates the sum of rows into a global array of number of rows
+// elements. First, partial reduction is computed inside each SG, then atomic
+// add is used to reduce between SG leaders
 
 #include <iostream>
 #include <sycl/sycl.hpp>
@@ -19,6 +21,6 @@
 using namespace sycl;
 using namespace sycl::ext::oneapi::experimental::matrix;
 
-#define SG_SZ 16
+#define SG_SZ 8
 
-#include "joint_matrix_int8_vnni_impl.hpp"
+#include "../element_wise_irreg_sum_rows_impl.hpp"
