+++ SYCL/Matrix/XMX8/element_wise_ops.cpp
@@ -1,17 +1,15 @@
-//==-------- joint_matrix_bf16_vnni.cpp  - DPC++ joint_matrix---------------==//
+//==----------- element_wise_ops.cpp  - DPC++ joint_matrix------------- ----==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
-// REQUIRES: matrix
+// REQUIRES: matrix-xmx8
 
 // RUN: %clangxx -fsycl %s -o %t.out
 // RUN: %CPU_RUN_PLACEHOLDER %t.out
 // RUN: %GPU_RUN_PLACEHOLDER %t.out
-
-// XFAIL: gpu
 
 #include <iostream>
 #include <sycl/sycl.hpp>
@@ -19,6 +17,6 @@
 using namespace sycl;
 using namespace sycl::ext::oneapi::experimental::matrix;
 
-#define SG_SZ 16
+#define SG_SZ 8
 
-#include "joint_matrix_int8_vnni_impl.hpp"
+#include "../element_wise_ops_impl.hpp"
