+++ SYCL/Matrix/XMX8/joint_matrix_bfloat16_use.cpp
@@ -1,13 +1,13 @@
-//==-------- joint_matrix_bf16_vnni.cpp  - DPC++ joint_matrix---------------==//
+//==-------- joint_matrix_bfloat16_use.cpp  - DPC++ joint_matrix------------==//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //
 //===----------------------------------------------------------------------===//
-// REQUIRES: matrix
+// REQUIRES: matrix-xmx8
 
-// RUN: %clangxx -fsycl %s -o %t.out
+// RUN: %clangxx -fsycl %s -o %t.out -DSYCL_EXT_ONEAPI_MATRIX_VERSION=2
 // RUN: %CPU_RUN_PLACEHOLDER %t.out
 // RUN: %GPU_RUN_PLACEHOLDER %t.out
 
@@ -16,9 +16,9 @@
 #include <iostream>
 #include <sycl/sycl.hpp>
 
-using namespace sycl;
 using namespace sycl::ext::oneapi::experimental::matrix;
+using bfloat16 = sycl::ext::oneapi::experimental::bfloat16;
 
-#define SG_SZ 16
+#define SG_SZ 8
 
-#include "joint_matrix_int8_vnni_impl.hpp"
+#include "../joint_matrix_bfloat16_use_impl.hpp"
